Modularize FAAValign and support Python3
The FAAValign.py script has been broken up into three files
each covering different aspects of the program. The file
cmudictionary.py handles operations related to the dictionary.
The file transcriptprocessor.py handles the processing of the
input tsv file. The main module, aligner.py, handles the
aspects related to force alignment and serves as a
wrapper for interfacing with the previous two modules.

In the process of modularizing the code, it has been
rewritten to run under Python3 since Python2 is no longer
supported. The new modules have been linted to comply
with PEP standards, and the linter configuration is included
as .pylintrc in the root directory.