Don't print git diff output if not in git repo
When running the aligner, the logger checks for any changes in the
repo and prints the differences to the log. If there's no repo then
the error output gets printed to the terminal which can be confusing
for users. This commit first checks to see if a git repo exists, and
skips the process if not.