Add cmu_dict... test and fix write error
This commit adds some testing for the cmudictionary.py module.

In testing cmudictionary.CMU_Dictionary.add_dictionary_entries()
a bug was uncovered in .check_transcription() which caused the
output file to have more spaces than necessary. The cause was that
function did not split the phone string when it was supposed to.
This commit fixes the error and includes the test to prevent regression.