Fix unintended overwrite of add_dict
An error in cmudictionary.py caused add_dict to be overwritten and
change type when a transcribed word was not in the cmu dictionary and
had not yet been added to the output dictionary. This led to an obvious
type error and a not obvious failure to add custom dictionaries. This
commit fixes the issue by adding the given word as a key to add_dict
and adds a test to prevent regressions.

Resolves JoFrhwld/FAVE#59