Merge with upstream/dev
squashed commit of:

ipa column

aeh to ae:

working on docs

markdown depend

ipa column

aeh to ae:

poetry lock woes

move from rst to myst

conversion cleanup

Add myst-parser to docs/requirements.txt for build

The Github workflow uses the requirements.txt to figure out what
dependencies are needed for the documentation build.

Migrate documentation from ReStructured Text to MarkDown (#1)

* readme

* ipa column

* aeh to ae:

* working on docs

* markdown depend

* Update deps, expand author info in toml

(cherry picked from commit 6bc72a49896c6a6708bda7a411e9114034ce6434)

* Docs: write migration guide and doc updates

Update documentation for FAVE 2:
 * Include warning for old users on reading migration guide.
 * Write migration guide with FAQ
 * Simplify TOC code
 * Add index.rst for code and usage sections of docs
 * Add sphinx-project module into conf to make doc updates easier

* Add cmu_dict... test and fix write error

This commit adds some testing for the cmudictionary.py module.

In testing cmudictionary.CMU_Dictionary.add_dictionary_entries()
a bug was uncovered in .check_transcription() which caused the
output file to have more spaces than necessary. The cause was that
function did not split the phone string when it was supposed to.
This commit fixes the error and includes the test to prevent regression.

* Fix unintended overwrite of add_dict

An error in cmudictionary.py caused add_dict to be overwritten and
change type when a transcribed word was not in the cmu dictionary and
had not yet been added to the output dictionary. This led to an obvious
type error and a not obvious failure to add custom dictionaries. This
commit fixes the issue by adding the given word as a key to add_dict
and adds a test to prevent regressions.

Resolves JoFrhwld/FAVE#59

* Squashed commit of the following:

commit 6bc72a49896c6a6708bda7a411e9114034ce6434
Author: Christian Brickhouse <brickhouse@stanford.edu>
Date:   Wed Aug 17 23:10:36 2022 -0700

    Update deps, expand author info in toml

commit 694ba50f87a9c9ca5202ed19493319c0f162ea90
Author: Christian Brickhouse <brickhouse@stanford.edu>
Date:   Wed Aug 17 23:09:34 2022 -0700

    Bump version to 2.0.2-dev

commit 716f95e78c6e5092e8204ebab6b16911f876af36
Author: Christian Brickhouse <brickhouse@stanforf.edu>
Date:   Wed Aug 17 22:43:25 2022 -0700

    Fix unintended overwrite of add_dict

    An error in cmudictionary.py caused add_dict to be overwritten and
    change type when a transcribed word was not in the cmu dictionary and
    had not yet been added to the output dictionary. This led to an obvious
    type error and a not obvious failure to add custom dictionaries. This
    commit fixes the issue by adding the given word as a key to add_dict
    and adds a test to prevent regressions.

    Resolves JoFrhwld/FAVE#59

commit 40503a6d2f8735cf9516df1c8a2266c9daed0f97
Author: Christian Brickhouse <brickhouse@stanforf.edu>
Date:   Wed Aug 17 22:31:34 2022 -0700

    Add cmu_dict... test and fix write error

    This commit adds some testing for the cmudictionary.py module.

    In testing cmudictionary.CMU_Dictionary.add_dictionary_entries()
    a bug was uncovered in .check_transcription() which caused the
    output file to have more spaces than necessary. The cause was that
    function did not split the phone string when it was supposed to.
    This commit fixes the error and includes the test to prevent regression.

* not appropriate to log git versions anymore

* fixing a copy-paste error from elsewhere

* fixing the tqdm pbar update

* updating

* ipa column

* aeh to ae:

* poetry lock woes

* move from rst to myst

* conversion cleanup

* Add myst-parser to docs/requirements.txt for build

The Github workflow uses the requirements.txt to figure out what
dependencies are needed for the documentation build.

Co-authored-by: Christian Brickhouse <brickhouse@stanford.edu>
Co-authored-by: Christian Brickhouse <brickhouse@stanforf.edu>

Update deps, expand author info in toml